+++ tests/models/test_press_summaries.py
@@ -110,6 +110,7 @@ class TestPressSummaryValidation:
         assert press_summary.validation_failure_messages == sorted(
             [
                 "This document failed to parse",
+                "This press summary does not have a unique content hash",
                 "This press summary is currently parked at a temporary URI",
                 "This press summary is currently on hold",
                 "This press summary has no name",